fix(ci): take the summary verdict from the job results
Counting summary files cannot see a job that produced none. One matrix
cell killed by its timeout still left two files, so the dead cell was
absent from the table and from the totals, and the run was reported as a
success. A status the job wrote verbatim had the same hole: `skipped`
matched no counter and passed silently.

The verdict now reads the two job results, which are reported whatever a
job managed to upload, and anything other than success fails the run and
is named in the summary. The file counts describe the table only.

Guard every published copy, not the module alone. The two meta-schemas
are generated by the same step and served at addresses that other tools
resolve, so a copy left behind is the same defect. A missing copy now
fails with the same instruction as a stale one.
